Function 3 completely working.
My charReplace function is working great. At first I had to go into the
function and edit in which character I wanted switched. I did research
and found the RegExp function and implemented that into the function in
order to declare which character I wanted to switch when calling the
function.